refactor(swarm-node): extract nat platform module to collapse wasm cfg gates
NAT traversal (AutoNAT v2, UPnP) and LAN discovery (mDNS) move out of
base.rs into a platform module: node/nat.rs composes them as a single
NatBehaviour sub-behaviour with one NatEvent enum and one dispatcher,
and node/nat_wasm.rs is the browser sibling with identical item names
and signatures over a no-op dummy behaviour and an uninhabited event.

The client and bootnode composites now carry one nat field, one event
variant, one From impl, and one match arm on both targets, removing
every NAT-related target cfg from client.rs and base.rs and the
duplicated field/From/match block from bootnode.rs. Handler bodies
move verbatim; toggles are driven by the same config flags and the
composite field order is unchanged, so native behaviour is identical.

docs/agents/wasm.md documents the approach as Pattern C (platform
module) with hard rules against cfg-gating struct fields or enum
variants.
