+++ docs/agents/wasm.md
@@ -10,7 +10,7 @@ A wasm client unlocks light-client embeddings in dapps, indexer UIs, and the Nex
 
 - Already no_std capable with a `default = ["std"]` feature: `vertex-swarm-primitives`, `vertex-swarm-spec`, `vertex-swarm-forks`, `vertex-swarm-api`.
 - The peer stack (`vertex-swarm-peer`, `vertex-swarm-peer-score`, `vertex-swarm-peer-manager` plus their `vertex-net-peer-*` and `vertex-net-local` deps) builds for `wasm32-unknown-unknown` and CI enforces it (the `wasm` job in `.github/workflows/unit.yml`). The build currently requires a nightly toolchain: `nectar-primitives` pulls `wasm-bindgen-rayon` on wasm32, which needs the unstable `atomics` target feature that stable rustc does not expose. The required rustflags (`+atomics,+bulk-memory,+mutable-globals` and the `getrandom_backend="wasm_js"` cfg) live in `.cargo/config.toml` under `[target.wasm32-unknown-unknown]`, mirroring nectar's own config. Run it with `cargo +nightly build --target wasm32-unknown-unknown -p vertex-util-runtime -p vertex-swarm-peer-score -p vertex-swarm-peer-manager -p vertex-swarm-identity`. Making this buildable on stable is upstream work in nectar (gate the `wasm-bindgen-rayon` pull behind a feature).
-- The full client cone (`vertex-swarm-node`) builds for `wasm32-unknown-unknown` and CI enforces it (the `wasm` job builds `-p vertex-swarm-node`). This pulls the topology composite behaviour and every `/swarm/...` wire protocol. The three native-only blockers in the topology cone were resolved: `vertex-observability` was split so the `axum` metrics server is a native-default feature, `vertex-net-dnsaddr` bootnode resolution is cfg-gated to native with a `vertex-net-dnsaddr-doh` (DNS-over-HTTPS) sibling on wasm, and the `if-watch` netdev interface watcher is cfg-gated to native with a no-op wasm sibling. Topology and the wire crates pull a trimmed wasm libp2p (no tcp/dns/mdns/upnp). A headless mainnet connection cannot run in CI (it needs a browser and the network); the compile is the CI-enforced proof, and the live connect is exercised by the browser demo.
+- The full client cone (`vertex-swarm-node`) builds for `wasm32-unknown-unknown` and CI enforces it (the `wasm` job builds `-p vertex-swarm-node`). This pulls the topology composite behaviour and every `/swarm/...` wire protocol. The three native-only blockers in the topology cone were resolved: `vertex-observability` was split so the `axum` metrics server is a native-default feature, `vertex-net-dnsaddr` bootnode resolution is cfg-gated to native with a `vertex-net-dnsaddr-doh` (DNS-over-HTTPS) sibling on wasm, and the `if-watch` netdev interface watcher is cfg-gated to native with a no-op wasm sibling. Topology and the wire crates pull a trimmed wasm libp2p (no tcp/dns/mdns/upnp). A headless mainnet connection cannot run in CI (it needs a browser and the network); the compile is the CI-enforced proof, and the live connect is exercised by the browser demo. Inside `vertex-swarm-node`, NAT traversal (AutoNAT v2, UPnP) and LAN discovery (mDNS) live in a platform module (`node/nat.rs` with the `node/nat_wasm.rs` sibling, Pattern C below): the node composites carry a single `nat` field on both targets, and the wasm sibling is a no-op behaviour with an uninhabited event enum.
 - Time and randomness for the wasm cone live in one cfg-gated home, `vertex-util-runtime` (`crates/util/runtime`). Its `time` module re-exports the `web-time` clock types (a `std::time` re-export on native, browser clock on wasm32) and adds deduplicated Unix-timestamp helpers (`now_unix_secs`, `now_unix_millis`, `now_unix_nanos`, `now`); its `rand` module is a getrandom-backed facade (`fill_bytes`, `crypto_rng`, `non_crypto_rng`, `try_*`) standardized on rand 0.9 / getrandom 0.3.4. There are no thread-local RNGs. Reach for `vertex_util_runtime::time` and `vertex_util_runtime::rand` rather than importing `web-time`, `getrandom`, or `rand` directly. The single intentional `web-time` dependency is `vertex-util-runtime` itself; everything else routes through the facade.
 - Identity in the wasm cone is ephemeral by design. The keystore (`vertex-swarm-identity` on-disk keystore) is native-only; a wasm client constructs an in-memory `Identity::random()` per session rather than persisting a key. There is no wasm keystore and none is planned.
 - Nectar primitives (`nectar-primitives`, `nectar-mantaray`, `nectar-postage`) are the upstream wasm-friendly layer. The proof-of-concept `crates/wasm-demo` lives in nectar.
@@ -53,7 +53,7 @@ The grey zone (must be cfg-gated):
 
 ### cfg-gating rules
 
-Pick one of two patterns and stick to it per crate. Mixing both is the source of all wasm build breakages.
+Pick one of these patterns and stick to it per crate. Mixing them is the source of all wasm build breakages.
 
 **Pattern A: `target_arch = "wasm32"`** for transport-level code where the choice is platform-determined and the consumer should not have to opt in. Use this for runtime selection (`tokio::spawn` vs `wasm_bindgen_futures::spawn_local`), getrandom backend, and similar. Example:
 
@@ -67,8 +67,12 @@ pub fn spawn<F: Future + 'static>(f: F) { wasm_bindgen_futures::spawn_local(f); 
 
 **Pattern B: cargo features** (`std`, `wasm`, `native`) for capability-level code where a single crate offers two implementations and the consumer crate picks one. Use this for storage backends (`vertex-storage-redb` is feature-gated out in wasm builds; an in-memory `Database` impl is feature-gated in) and observability sinks. Default features should produce the native build; the wasm build sets `default-features = false` and selects the wasm feature.
 
-Rules that apply to both:
+**Pattern C: platform module.** When a single target cfg would touch three or more items in one file, or would gate a struct field, extract a sibling module selected with `#[cfg_attr(target_arch = "wasm32", path = "foo_wasm.rs")] mod foo;`. Both files export the same item names with identical signatures so every call site compiles unchanged on both targets. Models: `vertex-swarm-topology` (`tasks.rs`/`tasks_wasm.rs`) and `vertex-swarm-node` (`nat.rs`/`nat_wasm.rs`).
 
+Rules that apply to all of them:
+
+- Never cfg-gate struct fields or enum variants. Wrap the divergent part in a platform type (Pattern C) with a uniform constructor. For a `NetworkBehaviour` composite, the wasm sibling is a one-field wrapper around `libp2p::swarm::dummy::Behaviour` whose `to_swarm` event is an uninhabited enum; the shared composite keeps one field, one event variant, one `From` impl, one match arm. Do not derive `NetworkBehaviour` on a zero-field struct.
+- Sibling signatures are byte-identical: same parameters, same (or alias-equivalent) return types. The wasm side accepts and ignores arguments rather than changing arity.
 - No `#[cfg(target_arch = "wasm32")]` inside a function body. Hoist to a function-level cfg and provide two implementations. This keeps the code path obvious to a reader and gives clippy a clean signal.
 - `#[cfg(not(target_arch = "wasm32"))]` blocks must have a matching wasm sibling. If the wasm path is "not supported", make the function return `Result` with a documented error variant, not a panic.
 - New crates added to the wasm cone must include `wasm32-unknown-unknown` in their `[lints]` or `[lints.target.'cfg(target_arch = "wasm32")']` table so clippy runs on both targets in CI.
